+++ src/main/java/com/example/airtraffic/repository/RunwayRepo.java
@@ -3,5 +3,5 @@ package com.example.airtraffic.repository;
 import com.example.airtraffic.model.Runway;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RunwayRepo extends JpaRepository<Runway,Long> {
+public interface RunwayRepo extends JpaRepository<Runway,Integer> {
 }
